DAS-518 refactor a bit and add tests.
